feat(Hands On Generator): Implemented the Simple Dungeon Generator
I call it simple because after reading the chapter (and reimplementing it with minimum reference).
I realized there were many types of algorithms to do it, among which one using Grammatical Graph Evolution (which I want to try implementing).

For that reason I implemented mine as generic over the dungeon generators so that when the time comes,
I can easily swap the two. I can also have various types of generators to have various flavors of dungeons.

That's a stretch goal though.